Introduce PreAnalysisHooks
PreAnalysisHooks are operations which can be run by the ZetaSQLToolkitAnalyzer after parsing a statement but before analyzing it. When hooks are run, they can perform validations or apply changes to the catalog as required.

Some hooks are always used by the analyzer, such as the
VariableDeclarationHook or the SingleAssignment hook. Catalogs can
define a set of hooks that should be applied when analyzing with them as
well.