Make ColumnLineageExtractor accept only concrete statement types
Previously, the API for ColumnLineageExtractor used the method
::extractColumnLevelLineage(ResolvedStatement). Since introducing
support for ResolvedQueryStmts, which needs specifying an output table
separately; maintaining the generic ResolvedStatement API required
making it confusing, since it would optionally need to accept an output
table.

This makes it so that teams building lineage applications need to
explicitly determine the statements they support and call the
corresponding ::extractColumnLevelLineage() method. Such as
::extractColumnLevelLineage(ResolvedInsertStmt) or
::extractColumnLevelLineage(ResolvedQueryStmt, String).